move cds plugin code into src to have more exports for testing
